examples/onnxrt/body_analysis/onnx_model_zoo/emotion_ferplus/quantization/ptq/requirements.txt
onnx==1.11.0
onnxruntime==1.10.0
onnxruntime_extensions
pillow>=8.2.0 # not directly required, pinned by Snyk to avoid a vulnerability
pandas